refactor: MatchCount 패키지 이동
