+++ src/main/java/lotto/domain/MatchCount.java
@@ -1,4 +1,4 @@
-package lotto.view;
+package lotto.domain;
 
 import java.util.Arrays;
 import java.util.Map;
