+++ src/test/java/lotto/domain/MatchCountTest.java
@@ -1,4 +1,4 @@
-package lotto.view;
+package lotto.domain;
 
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
